% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrainSub.R
\docType{data}
\name{mbrainSub}
\alias{mbrainSub}
\title{Subset of 1k Brain Cells from an E18 Mouse}
\format{
An object of class \code{"dgCMatrix"}.
}
\source{
\href{http://cf.10xgenomics.com/samples/cell-exp/2.1.0/neurons_900}{1k Brain Cells from an E18 Mouse}
}
\usage{
data(mbrainSub)
}
\description{
1k Brain Cells from an E18 Mouse is a public dataset from 10X Genomics.
This subset is the first 50,000 barcodes of original data.
}
\examples{
data(mbrainSub)
str(mbrainSub)

}
\keyword{datasets}
