% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read10X.R
\name{Read10X}
\alias{Read10X}
\title{Read 10X output data}
\usage{
Read10X(dir, row.name = "name", meta = F)
}
\arguments{
\item{dir}{The directory of 10X output data. For Cell Ranger version <3, 
directory should include three files: barcodes.tsv, genes.tsv, matrix.mtx. 
For Cell Ranger version >=3, directory should include three 
files: barcodes.tsv.gz, features.tsv.gz, matrix.mtx.gz.}

\item{row.name}{Specify either using gene names (\code{row.name = "name"}) or 
gene Ensembl IDs (\code{row.name = "id"}) as row names of the count matrix. 
Default is \code{row.name = "name"}.}

\item{meta}{Logical. If \code{T}, returns a list containing both the count matrix
and metadata of genes (features). Metadata includes feature names, IDs and other 
additional information depending on Cell Ranger output. If \code{F} (default),
only returns the count matrix.}
}
\value{
If \code{meta = T}, returns a list of two elements: a "dgCMatrix" 
sparse matrix containing expression counts and a data frame containing metadata of 
genes (features). For the count matrix, each row is a gene (feature) and 
each column is a barcode.  If \code{meta = F}, only returns the count matrix.
}
\description{
One-line handy function to read 10X Cell Ranger output data. Works under both old 
(<3) and new (>=3) Cell Ranger version.
}
\examples{
\dontrun{
data_dir <- 'path/to/data/directory'

#count matrix
data_mat <- Read10X(data.dir = data_dir)
str(data_mat)

#a list including count matrix and feature metadata
data_list <- Read10X(data.dir = data_dir, meta = T)
str(data_list)
}

}
